COMPLETED - task ccf314: [PT2QC]Null pointer exception thrown when added link attachment in PT http://ccf.open.collab.net/servlets/Scarab?id=ccf314 Only delete attachment file if it has been created before
git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk/core@2607 6771e1e2-0f46-0410-b5d7-d846ff89f743

+++ src/plugins/HPQC/v90/com/collabnet/ccf/pi/qc/v90/QCWriter.java
@@ -783,10 +783,12 @@ public class QCWriter extends AbstractWriter<IConnection> implements
 					.setErrorCode(GenericArtifact.ERROR_EXTERNAL_SYSTEM_CONNECTION);
 			throw new CCFRuntimeException(cause, e);
 		} finally {
-			boolean deletingSuccess = attachmentFile.delete();
-			if (!deletingSuccess) {
-				log.warn("Could not delete the attachment file "
-						+ attachmentFile.getAbsolutePath());
+			if (attachmentFile != null) {
+				boolean deletingSuccess = attachmentFile.delete();
+				if (!deletingSuccess) {
+					log.warn("Could not delete the attachment file "
+							+ attachmentFile.getAbsolutePath());
+				}
 			}
 			disconnect(connection);
 		}
